+++ changelog.md
@@ -17,6 +17,8 @@ Codion Change Log
 ### is.codion.framework.domain
 - KeyGenerator.Identity removed.
 - AttributeDefinition.Builder.valueRange(), minimumValue() and maximumValue() renamed range(), minimum() and maximum().
+### is.codion.plugin.flatlaf.intellij
+- Catppuccin themes added.
 
 ## 0.18.48
 ### is.codion.swing.common.model